requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-qdrant
langchain-groq
langchain-openai
langchain-text-splitters
qdrant-client
flashrank
ragas
datasets
sentence-transformers
pydantic
pydantic-settings
requests
google-generativeai
langchain-google-genai
